Implemented a collection of CheckOutItems
Created an implementation to store scanned items with
CheckOutItemCollection. This uses two structures internally, an array
for simple storage of al litems and a tally, to count the number of
each type of item, easier for price rules. The CheckOut class uses this
internally via the interface, allowing a different implementation to be
swapped out.

Unit tests added to cover various item list scenarios, including those
listed in the test spec. Removed the main.spec.ts as this was
duplicating test for CheckOut.spec.ts