Implemented Discount Price Rules and their collections
Created the idea of a IDiscount interface that each specific Discount
implements by requiring an interface to the checkout item collection.

So, a collection of discounts is created and then after all items are
scanned, the CheckOut passes the items scanned into the discount
collection, where each discount runs over the item collection to see if
its discount should be appliued. The total discounts are added up and
also their descriptions stored, so we can show all discounts that
applied.

Unit tests for each new class created. A new DiscountFactory created to
simplify testing and scenario building.

main scenarios script updated to run a number of scenarios given a set
of items scanned and discounts - which shows the total price and
discounts applied.