Implemented some fault tolerance, error reporting and observing
Implemented validation checks on Checkout items and discounts as they
are added to their collections. These throw specific errors with set
messages.

Created ErrorObserver which can be passed to the collections (a cleaner
way could be devised with more time) and outputs any errors thrown.
It also allows the caller to determine if an exception will halt
execution of the transaction or if the transaction should continue
on, ignoring bad input.

NOTE: could have added further validation on result of discount,
to make sure discounts are not > total prices.